# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valet Parking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginBody {
  code: string;
  /** Required for driver role */
  driverName?: string;
  /** Required for admin/driver role */
  eventId?: number;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  owner: "owner",
  admin: "admin",
  driver: "driver",
} as const;

export interface LoginResponse {
  role: LoginResponseRole;
  driverName?: string;
  eventId?: number;
  sessionToken: string;
}

export type SessionInfoRole =
  (typeof SessionInfoRole)[keyof typeof SessionInfoRole];

export const SessionInfoRole = {
  owner: "owner",
  admin: "admin",
  driver: "driver",
} as const;

export interface SessionInfo {
  role: SessionInfoRole;
  driverName?: string;
  eventId?: number;
}

export interface AppConfig {
  id: number;
  appName: string;
  updatedAt?: string;
}

export interface UpdateConfigBody {
  appName?: string;
}

export interface Event {
  id: number;
  name: string;
  eventDate: string;
  isActive: boolean;
  createdAt: string;
  activeShiftId?: number | null;
  hasOpenShift: boolean;
}

export interface CreateEventBody {
  name: string;
  eventDate: string;
}

export interface UpdateEventBody {
  name?: string;
  eventDate?: string;
  isActive?: boolean;
}

export type EventStatsByLocationItem = {
  locationId: number;
  locationName: string;
  count: number;
  capacity: number;
};

export interface EventStats {
  totalActive: number;
  totalDelivered: number;
  totalInTransit: number;
  totalRelocated: number;
  byLocation: EventStatsByLocationItem[];
}

export interface Shift {
  id: number;
  eventId: number;
  label: string;
  openedBy: string;
  openedAt: string;
  closedAt?: string | null;
  isOpen: boolean;
}

export interface CreateShiftBody {
  eventId: number;
  label: string;
  openedBy: string;
}

export interface UpdateShiftBody {
  label?: string;
  closedAt?: string;
  isOpen?: boolean;
}

export type ValetTicketKeyLocation =
  (typeof ValetTicketKeyLocation)[keyof typeof ValetTicketKeyLocation];

export const ValetTicketKeyLocation = {
  drawer: "drawer",
  board: "board",
  not_found: "not_found",
  with_owner: "with_owner",
} as const;

export type ValetTicketStatus =
  (typeof ValetTicketStatus)[keyof typeof ValetTicketStatus];

export const ValetTicketStatus = {
  active: "active",
  in_transit: "in_transit",
  delivered: "delivered",
  relocated: "relocated",
} as const;

export type ValetTicketVehicleDamagesItem =
  (typeof ValetTicketVehicleDamagesItem)[keyof typeof ValetTicketVehicleDamagesItem];

export const ValetTicketVehicleDamagesItem = {
  front_door: "front_door",
  rear_bumper: "rear_bumper",
  front_bumper: "front_bumper",
  rear_light: "rear_light",
  front_light: "front_light",
  hood: "hood",
} as const;

export interface ValetTicket {
  id: number;
  eventId: number;
  shiftId: number;
  valetNumber: string;
  driverName: string;
  keyLocation: ValetTicketKeyLocation;
  parkingLocationId?: number | null;
  parkingLocationName?: string | null;
  status: ValetTicketStatus;
  vehicleDamages: ValetTicketVehicleDamagesItem[];
  notes?: string | null;
  relocatedToLocationId?: number | null;
  relocatedToLocationName?: string | null;
  vehicleColor?: string | null;
  vehicleBrand?: string | null;
  licensePlate?: string | null;
  createdAt: string;
  deliveredAt?: string | null;
}

export type CreateTicketBodyKeyLocation =
  (typeof CreateTicketBodyKeyLocation)[keyof typeof CreateTicketBodyKeyLocation];

export const CreateTicketBodyKeyLocation = {
  drawer: "drawer",
  board: "board",
  not_found: "not_found",
  with_owner: "with_owner",
} as const;

export type CreateTicketBodyVehicleDamagesItem =
  (typeof CreateTicketBodyVehicleDamagesItem)[keyof typeof CreateTicketBodyVehicleDamagesItem];

export const CreateTicketBodyVehicleDamagesItem = {
  front_door: "front_door",
  rear_bumper: "rear_bumper",
  front_bumper: "front_bumper",
  rear_light: "rear_light",
  front_light: "front_light",
  hood: "hood",
} as const;

export interface CreateTicketBody {
  eventId: number;
  shiftId: number;
  valetNumber: string;
  driverName: string;
  keyLocation: CreateTicketBodyKeyLocation;
  parkingLocationId: number;
  vehicleDamages?: CreateTicketBodyVehicleDamagesItem[];
  notes?: string;
  vehicleColor?: string;
  vehicleBrand?: string;
  licensePlate?: string;
}

export type UpdateTicketBodyStatus =
  (typeof UpdateTicketBodyStatus)[keyof typeof UpdateTicketBodyStatus];

export const UpdateTicketBodyStatus = {
  active: "active",
  in_transit: "in_transit",
  delivered: "delivered",
  relocated: "relocated",
} as const;

export type UpdateTicketBodyKeyLocation =
  (typeof UpdateTicketBodyKeyLocation)[keyof typeof UpdateTicketBodyKeyLocation];

export const UpdateTicketBodyKeyLocation = {
  drawer: "drawer",
  board: "board",
  not_found: "not_found",
  with_owner: "with_owner",
} as const;

export type UpdateTicketBodyVehicleDamagesItem =
  (typeof UpdateTicketBodyVehicleDamagesItem)[keyof typeof UpdateTicketBodyVehicleDamagesItem];

export const UpdateTicketBodyVehicleDamagesItem = {
  front_door: "front_door",
  rear_bumper: "rear_bumper",
  front_bumper: "front_bumper",
  rear_light: "rear_light",
  front_light: "front_light",
  hood: "hood",
} as const;

export interface UpdateTicketBody {
  status?: UpdateTicketBodyStatus;
  keyLocation?: UpdateTicketBodyKeyLocation;
  parkingLocationId?: number;
  relocatedToLocationId?: number;
  vehicleDamages?: UpdateTicketBodyVehicleDamagesItem[];
  notes?: string;
}

export interface ParkingLocation {
  id: number;
  name: string;
  capacity: number;
  currentCount: number;
  colorHex: string;
  isActive: boolean;
}

export interface CreateParkingLocationBody {
  name: string;
  capacity: number;
  colorHex: string;
}

export interface UpdateParkingLocationBody {
  name?: string;
  capacity?: number;
  colorHex?: string;
  isActive?: boolean;
}

export type AccessCodeRole =
  (typeof AccessCodeRole)[keyof typeof AccessCodeRole];

export const AccessCodeRole = {
  owner: "owner",
  admin: "admin",
  driver: "driver",
} as const;

export interface AccessCode {
  id: number;
  eventId?: number | null;
  code: string;
  role: AccessCodeRole;
  label: string;
  isActive: boolean;
  expiresAfterShiftClose: boolean;
  createdAt: string;
}

export type CreateAccessCodeBodyRole =
  (typeof CreateAccessCodeBodyRole)[keyof typeof CreateAccessCodeBodyRole];

export const CreateAccessCodeBodyRole = {
  admin: "admin",
  driver: "driver",
} as const;

export interface CreateAccessCodeBody {
  eventId?: number;
  code: string;
  role: CreateAccessCodeBodyRole;
  label: string;
  expiresAfterShiftClose?: boolean;
}

export type ListShiftsParams = {
  eventId: number;
};

export type ListTicketsParams = {
  eventId: number;
  shiftId?: number;
  status?: ListTicketsStatus;
  search?: string;
};

export type ListTicketsStatus =
  (typeof ListTicketsStatus)[keyof typeof ListTicketsStatus];

export const ListTicketsStatus = {
  active: "active",
  in_transit: "in_transit",
  delivered: "delivered",
  relocated: "relocated",
} as const;

export type ListTicketHistoryParams = {
  eventId: number;
  from?: string;
  to?: string;
};

export type SearchTicketByValetNumberParams = {
  eventId: number;
};

export type ListParkingLocationsParams = {
  eventId?: number;
};

export type ListAccessCodesParams = {
  eventId: number;
};

export interface BlockedIp {
  ip: string;
  count: number;
  lastAttempt: string;
}

export interface BulkGenerateCodesBody {
  eventId: number;
  driverCount: number;
  adminCount: number;
}