# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Valet Parking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Not authenticated
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (owner only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (owner only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /events/{eventId}/clear-history:
    post:
      operationId: clearEventHistory
      tags: [events]
      summary: Clear event ticket history (owner only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cleared
  /events/{eventId}/stats:
    get:
      operationId: getEventStats
      tags: [events]
      summary: Get event statistics (active tickets, delivered, by location)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStats"
...
  /access-codes:
    get:
      operationId: listAccessCodes
      tags: [access-codes]
      summary: List access codes for an event (owner only)
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessCode"
    post:
      operationId: createAccessCode
      tags: [access-codes]
      summary: Create an access code (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccessCodeBody"
      responses:
        "201":
          description: Created code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
  /access-codes/{codeId}:
    delete:
      operationId: deleteAccessCode
      tags: [access-codes]
      summary: Delete/revoke an access code (owner only)
      parameters:
        - name: codeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    LoginBody:
      type: object
      properties:
        code:
          type: string
        driverName:
          type: string
          description: Required for driver role
        eventId:
          type: integer
          description: Required for admin/driver role
      required: [code]
    LoginResponse:
      type: object
      properties:
        role:
          type: string
          enum: [owner, admin, driver]
        driverName:
          type: string
        eventId:
          type: integer
        sessionToken:
          type: string
      required: [role, sessionToken]
    SessionInfo:
      type: object
      properties:
        role:
          type: string
          enum: [owner, admin, driver]
        driverName:
          type: string
        eventId:
          type: integer
      required: [role]
...
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        eventDate:
          type: string
          format: date
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        activeShiftId:
          type: integer
          nullable: true
        hasOpenShift:
          type: boolean
      required: [id, name, eventDate, isActive, createdAt, hasOpenShift]
    CreateEventBody:
      type: object
      properties:
        name:
          type: string
        eventDate:
          type: string
          format: date
      required: [name, eventDate]
    UpdateEventBody:
      type: object
      properties:
        name:
          type: string
        eventDate:
          type: string
          format: date
        isActive:
          type: boolean
    EventStats:
      type: object
      properties:
        totalActive:
          type: integer
        totalDelivered:
          type: integer
        totalInTransit:
          type: integer
        totalRelocated:
          type: integer
        byLocation:
          type: array
          items:
            type: object
            properties:
              locationId:
                type: integer
              locationName:
                type: string
              count:
                type: integer
              capacity:
                type: integer
            required: [locationId, locationName, count, capacity]
      required: [totalActive, totalDelivered, totalInTransit, totalRelocated, byLocation]
...
    AccessCode:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
          nullable: true
        code:
          type: string
        role:
          type: string
          enum: [owner, admin, driver]
        label:
          type: string
        isActive:
          type: boolean
        expiresAfterShiftClose:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, code, role, label, isActive, expiresAfterShiftClose, createdAt]
    CreateAccessCodeBody:
      type: object
      properties:
        eventId:
          type: integer
        code:
          type: string
        role:
          type: string
          enum: [admin, driver]
        label:
          type: string
        expiresAfterShiftClose:
          type: boolean
      required: [code, role, label]